pyproject.toml
[tool.poetry]
name = "mariafm"
version = "0.1.0"
description = "Одностраничный сайт на Django для радиостанции mariafm"
authors = ["kotelnikov-vladislav <vladislav.kotelnikov@internet.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-vite = "^3.0.5"
whitenoise = "^6.7.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
